add canvas
新增canvas js,html,css
修改.gitgnore-.DS_Store
