+++ app/controllers/pages_controller.rb
@@ -3,4 +3,9 @@
 class PagesController < ApplicationController
   layout "index"
   def home; end
+  def canvas
+    render layout: false
+  
+  end
+
 end
