+++ app/javascript/canvas/index.js
@@ -1,0 +1,2 @@
+import "./canvas.css"
+import "./canvas.js"
