+++ app/javascript/packs/application.js
@@ -9,4 +9,5 @@ ActiveStorage.start()
 
 import "controllers"
 import "stylesheets"
+import "canvas"
 import "../application.css"
