+++ app/views/pages/canvas.html.erb
@@ -4,8 +4,10 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="style.css" />
+    # <!-- <link rel="stylesheet" href="canvas.css" /> -->
     <title>Document</title>
+    <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
+    <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
   </head>
   <body>
     <section class="container">
@@ -24,6 +26,6 @@
         <canvas id="drawing-board"></canvas>
       </div>
     </section>
-    <script src="index.js"></script>
+    # <!-- <script src="canvas.js"></script> -->
   </body>
 </html>
