+++ config/routes.rb
@@ -8,6 +8,7 @@ Rails.application.routes.draw do
   end
 
   get 'pages/home'
+  get '/canvas', to: 'pages#canvas'
   root to: 'pages#home'
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
